+++ components/toolstab/impl/src/main/java/com/flipperdevices/toolstab/impl/api/ToolsDecomposeComponentImpl.kt
@@ -9,6 +9,7 @@ import com.flipperdevices.bottombar.handlers.ResetTabDecomposeHandler
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.deeplink.model.Deeplink
 import com.flipperdevices.nfc.mfkey32.api.MfKey32DecomposeComponent
+import com.flipperdevices.remotecontrols.api.RemoteControlsScreenDecomposeComponent
 import com.flipperdevices.toolstab.api.ToolsDecomposeComponent
 import com.flipperdevices.toolstab.impl.model.ToolsNavigationConfig
 import com.flipperdevices.toolstab.impl.model.toConfigStack
@@ -26,7 +27,8 @@ class ToolsDecomposeComponentImpl @AssistedInject constructor(
     @Assisted deeplink: Deeplink.BottomBar.ToolsTab?,
     @Assisted private val onBack: DecomposeOnBackParameter,
     private val hubMainFactory: ToolsMainScreenDecomposeComponentImpl.Factory,
-    private val mfKey32Factory: MfKey32DecomposeComponent.Factory
+    private val mfKey32Factory: MfKey32DecomposeComponent.Factory,
+    private val remoteControlsComponentFactory: RemoteControlsScreenDecomposeComponent.Factory,
 ) : ToolsDecomposeComponent<ToolsNavigationConfig>(),
     ComponentContext by componentContext,
     ResetTabDecomposeHandler {
@@ -48,6 +50,11 @@ class ToolsDecomposeComponentImpl @AssistedInject constructor(
         )
 
         ToolsNavigationConfig.MfKey32 -> mfKey32Factory(
+            componentContext = componentContext,
+            onBack = { navigation.popOr(onBack::invoke) }
+        )
+
+        ToolsNavigationConfig.RemoteControls -> remoteControlsComponentFactory(
             componentContext = componentContext,
             onBack = { navigation.popOr(onBack::invoke) }
         )