+++ components/toolstab/impl/src/main/java/com/flipperdevices/toolstab/impl/composable/ComposableHub.kt
@@ -6,11 +6,13 @@ import androidx.compose.ui.Modifier
 import com.flipperdevices.core.ui.ktx.OrangeAppBar
 import com.flipperdevices.toolstab.impl.R
 import com.flipperdevices.toolstab.impl.composable.elements.MifareClassicComposable
+import com.flipperdevices.toolstab.impl.composable.elements.RemoteControlsComposable
 
 @Composable
 fun ComposableHub(
     hasNotification: Boolean,
     onOpenMfKey32: () -> Unit,
+    onOpenRemoteControls: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     Column(modifier) {
@@ -20,6 +22,9 @@ fun ComposableHub(
         MifareClassicComposable(
             hasMfKey32Notification = hasNotification,
             onOpenMfKey32 = onOpenMfKey32
+        )
+        RemoteControlsComposable(
+            onOpenRemoteControls = onOpenRemoteControls
         )
     }
 }